requirements.txt
aiofiles==0.6.0
astroid==2.5.1
bcrypt==3.2.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.6
dnspython==2.1.0
ecdsa==0.14.1
email-validator==1.1.2
fastapi==0.63.0
gunicorn==20.0.4
h11==0.12.0
httptools==0.4.0
idna==3.1
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.1
pylint==2.7.2
python-jose==3.2.0
python-multipart==0.0.5
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.3.23
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
uvicorn==0.13.4
wrapt==1.12.1
